Organize core appreffinge definitions
