+++ packages/voictents-and-estinants-engine/src/core/engine-shell/appreffinge/input/inputAppreffinge.ts
@@ -1,6 +1,6 @@
 import { GenericLeftInputAppreffinge } from './leftInputAppreffinge';
-import { GenericRightInputAppreffinge } from './rightInputAppreffinge';
+import { GenericRightInputAppreffinge } from './right/rightInputAppreffinge';
 
-export type GenericAppreffinge2 =
+export type GenericInputAppreffinge2 =
   | GenericLeftInputAppreffinge
   | GenericRightInputAppreffinge;
