+++ packages/voictents-and-estinants-engine/src/core/engine-shell/appreffinge/input/leftInputAppreffinge.ts
@@ -1,4 +1,4 @@
-import { GenericLeftInputVicken } from '../vicken/leftInputVicken';
+import { GenericLeftInputVicken } from '../../vicken/leftInputVicken';
 
 export type LeftInputAppreffinge<
   TLeftInputVicken extends GenericLeftInputVicken,
