+++ packages/voictents-and-estinants-engine/src/core/engine-shell/appreffinge/input/right/framation.ts
@@ -1,7 +1,7 @@
-import { ZornTuple } from '../../../utilities/semantic-types/zorn';
-import { GenericIndexedHubblepup } from '../hubblepup/hubblepup';
-import { GenericLeftInputVicken } from '../vicken/leftInputVicken';
-import { GenericRightInputHubblepupTupleVicken } from '../vicken/rightInputVicken';
+import { ZornTuple } from '../../../../../utilities/semantic-types/zorn';
+import { GenericIndexedHubblepup } from '../../../hubblepup/hubblepup';
+import { GenericLeftInputVicken } from '../../../vicken/leftInputVicken';
+import { GenericRightInputHubblepupTupleVicken } from '../../../vicken/rightInputVicken';
 
 /**
  * A function that takes the leftmost Hubblepup of a Cology and outputs a Zorn tuple for each of the other members of the Cology
