+++ packages/voictents-and-estinants-engine/src/core/engine-shell/appreffinge/output/outputApreffinge.ts
@@ -1,4 +1,4 @@
-import { GenericOutputVicken } from '../vicken/outputVicken';
+import { GenericOutputVicken } from '../../vicken/outputVicken';
 
 export type OutputAppreffinge<TOutputVicken extends GenericOutputVicken> = {
   geppTuple: TOutputVicken['geppTuple'];
