+++ packages/voictents-and-estinants-engine/src/core/engine-shell/estinant/estinant.ts
@@ -1,8 +1,8 @@
 import { Tuple } from '../../../utilities/semantic-types/tuple';
 
-import { LeftInputAppreffinge } from '../appreffinge/leftInputAppreffinge';
-import { OutputAppreffinge } from '../appreffinge/outputApreffinge';
-import { InputVickenTupleToRightInputAppreffingeTuple } from '../appreffinge/rightInputAppreffinge';
+import { LeftInputAppreffinge } from '../appreffinge/input/leftInputAppreffinge';
+import { OutputAppreffinge } from '../appreffinge/output/outputApreffinge';
+import { InputVickenTupleToRightInputAppreffingeTuple } from '../appreffinge/input/right/rightInputAppreffinge';
 import { GenericLeftInputVicken } from '../vicken/leftInputVicken';
 import {
   GenericOutputVicken,
